requirements.txt
langgraph
yfinance
openai
pandas
numpy
matplotlib
python-dotenv 